backend/requirements.txt
google-generativeai>=0.3.2
python-dotenv>=1.0.1
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
pypdf<4.0,>=3.17.0
cryptography>=42.0.0
chromadb>=0.4.22
faiss-cpu>=1.7.4
pydantic>=2.6.1
Flask>=3.0.0
Flask-CORS>=4.0.0